third_party/blink/web_tests/accessibility/name-calc-button-inside-option.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div class="container">
  <div role="listbox">
    <div role="option" id="option1"><button>option 1</button></div>
    <div role="option" id="option2" aria-label="label 2">option 2</div>
  </div>
</div>

<script>
test(function(t) {
    var axOption1 = accessibilityController.accessibleElementById("option1");
    assert_equals(axOption1.name, "option 1");
    var axOption2 = accessibilityController.accessibleElementById("option2");
    assert_equals(axOption2.name, "label 2");
}, "ARIA listbox option with button inside");
</script>
